refactor: adicionar suporte para remoção de imagens e arquivos de ingressos na lógica de exclusão

+++ src/shared/helpers/deleteImages-and-files.ts
@@ -2,21 +2,27 @@ import { AwsService } from 'src/modules/aws/aws.service';
 import { extractFileName } from './extract-file-name';
 
 export async function deleteImagesAndFiles(
-  entities: { key: string; data: any[] }[],
+  entities: { key: string; data: any[]; type: 'images' | 'files' }[],
   awsService: AwsService,
 ) {
   await Promise.all(
-    entities.flatMap(({ key, data }) =>
+    entities.flatMap(({ key, data, type }) =>
       data.flatMap((entity) => {
-        const deleteImages = (entity.images ?? []).map((imageUrl) =>
-          awsService.delete(extractFileName(imageUrl, key), key),
-        );
+        if (type === 'images') {
+          const imageUrls = entity.imageUrls ?? entity.images ?? [];
+          return imageUrls.map((imageUrl) =>
+            awsService.delete(extractFileName(imageUrl, key), key),
+          );
+        }
 
-        const deleteFiles = (entity.files ?? []).map((fileUrl) =>
-          awsService.delete(extractFileName(fileUrl, key), key),
-        );
+        if (type === 'files') {
+          const fileUrls = entity.fileUrls ?? entity.files ?? [];
+          return fileUrls.map((fileUrl) =>
+            awsService.delete(extractFileName(fileUrl, key), key),
+          );
+        }
 
-        return [...deleteImages, ...deleteFiles];
+        return [];
       }),
     ),
   );
